Remove BackupJobService as the OS is scheduling its own backups
