+++ app/src/main/AndroidManifest.xml
@@ -55,10 +55,5 @@
             </intent-filter>
         </service>
 
-        <service
-            android:name=".service.backup.BackupJobService"
-            android:exported="false"
-            android:permission="android.permission.BIND_JOB_SERVICE" />
-
     </application>
 </manifest>
